Request current number of events
The XmlRpcProvider was separated from the BuildConfigResolver to allow the Notification Provider to use an XmlRpc authentication.
Event subscriptions were added because number of them are required for the request of current number of events.
Proxy for SummaryData object was added to provide an useful api for using it.
Credential will contain part of the RpcXml authentication info - Id of the user. It's required for a notification functionality.