Add cvs path detection, add cvs problem detection
Git path detection was taken from git vscode extension,
It looks for tf path in the vscode workspace configs and the environment paths.
Detection of active cvs was changed to transfer to the cvs provider the cvs path also.
Now it is shown what's the problem with cvs:
no cvs / incompatible version / not cvs rootPath / no staged changes